+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v3.0.4
 * Fixed: Infinite loop in PagingIterable if Elasticsearch returned vertices that wasn't saved in the database yet
+* Fixed: Fix Elasticsearch5 missing field exception when sorting on multiple indices
 
 # v3.0.3
 * Added: Added a hasAuthorization method to the Query class to allow searches for any element that uses an authorization string or strings.